refactor: drop localForage in favour of PouchDB
loginService already depends on PouchDB for its authentication plugin (which in
theory could be decoupled further in favour of $http), so use it as a local
storage wrapper instead of localForage.

This is implemented such that storage providers can be swapped out by injecting
a replacement `eha.login-service-adaptor.service` (with the casualty of
Java/enterprise-like inflection).

Note, coaxing the test suite to support angular-pouchdb (due to its use of
$q.when) was a gargantuan effort; somewhere down the line, ngMock changes the
normal behaviour of nested promises causing specs that test hit $q.when never to
resolve. This has cropped up time and again (ref the angular-pouchdb and
angular-localforage issue trackers). The only solution I've found so far is to
avoid using ngMock altogether. Besides the unfamiliarity to normal Angular
testing patterns, it does the job.

Within drudgery of fixing the tests, I swapped out Mocha (and Sinon) for Jasmine
for the simple fact that I'm more familiar with Jasmine in the context of unit
testing Angular and wanted to reduce some cognitive load. This conflates this
already huge diff somewhat, but it's late and I'm calling it a day.

Oh, and as Firefox was likely being used as a test launcher for it's native
Promise support and/or was left in as boilerplate from the Yeoman generator, I
swapped it out for PhantomJS (and made more use of $q) as it's easier to set up.

BREAKING CHANGE: localForage was dropped in favour of PouchDB. Credentials
previously managed by localForage are no longer accessible, but are otherwise
left intact. End-users will need to renter their credentials (likely through a
one-time dialog box).

Closes #2.